Refactored sample app
