+++ docs/guides/add-a-shared-dependency.md
@@ -1,0 +1,8 @@
+---
+order: 30
+label: Add a shared dependency
+---
+
+# Add a shared dependency
+
+
