+++ package.json
@@ -36,13 +36,13 @@
     "devDependencies": {
         "@changesets/changelog-github": "0.4.8",
         "@changesets/cli": "2.26.2",
-        "@typescript-eslint/parser": "6.7.0",
+        "@typescript-eslint/parser": "6.7.2",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/typescript-configs": "3.0.2",
         "cross-env": "7.0.3",
         "eslint": "8.49.0",
         "jest": "29.7.0",
-        "netlify-cli": "16.3.5",
+        "netlify-cli": "16.4.1",
         "retypeapp": "3.5.0",
         "ts-node": "10.9.1",
         "typescript": "5.2.2"
