+++ packages/core/package.json
@@ -35,7 +35,7 @@
         "react-dom": "*"
     },
     "devDependencies": {
-        "@types/react": "18.2.21",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/tsup-configs": "3.0.1",
