+++ packages/react-router/package.json
@@ -36,14 +36,14 @@
         "react-router-dom": "*"
     },
     "devDependencies": {
-        "@swc/core": "1.3.85",
+        "@swc/core": "1.3.86",
         "@swc/helpers": "0.5.2",
         "@swc/jest": "0.2.29",
         "@testing-library/react": "14.0.0",
         "@types/jest": "29.5.5",
-        "@types/react": "18.2.21",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
-        "@types/react-test-renderer": "18.0.1",
+        "@types/react-test-renderer": "18.0.2",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/swc-configs": "2.1.2",
         "@workleap/tsup-configs": "3.0.1",
