+++ packages/webpack-module-federation/package.json
@@ -43,18 +43,18 @@
         "webpack": ">=5.0.0"
     },
     "devDependencies": {
-        "@swc/core": "1.3.85",
+        "@swc/core": "1.3.86",
         "@swc/helpers": "0.5.2",
         "@swc/jest": "0.2.29",
         "@types/jest": "29.5.5",
-        "@types/node": "20.6.2",
-        "@types/react": "18.2.21",
+        "@types/node": "20.6.3",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/swc-configs": "2.1.2",
         "@workleap/tsup-configs": "3.0.1",
         "@workleap/typescript-configs": "3.0.2",
-        "@workleap/webpack-configs": "1.0.7",
+        "@workleap/webpack-configs": "1.0.8",
         "jest": "29.7.0",
         "react": "18.2.0",
         "react-dom": "18.2.0",
