--- sample/another-remote-module/src/App.tsx
@@ -1,30 +1,0 @@
-import { useAppRouter } from "@sample/shell";
-import { RouterProvider } from "react-router-dom";
-import { sessionManager } from "./session.ts";
-
-function DevHome() {
-    return (
-        <div>
-            <h2>Another remote module development home page</h2>
-            <p>Hey!</p>
-        </div>
-    );
-}
-
-export function App() {
-    const router = useAppRouter(sessionManager, {
-        managedRoutes: [
-            {
-                index: true,
-                element: <DevHome />
-            }
-        ]
-    });
-
-    return (
-        <RouterProvider
-            router={router}
-            fallbackElement={<div>Loading...</div>}
-        />
-    );
-}
