+++ sample/another-remote-module/webpack.dev.js
@@ -2,6 +2,7 @@
 
 import { defineDevRemoteModuleConfig } from "@squide/webpack-module-federation/defineConfig.js";
 import { defineDevConfig } from "@workleap/webpack-configs";
+import path from "node:path";
 import { swcConfig } from "./swc.dev.js";
 
 let config;
@@ -18,7 +19,9 @@ if (!process.env.LOCAL) {
         }
     });
 } else {
-    config = defineDevConfig(swcConfig);
+    config = defineDevConfig(swcConfig, {
+        entry: path.resolve("./src/dev/index.tsx")
+    });
 }
 
 export default config;
