+++ sample/host/package.json
@@ -14,16 +14,16 @@
         "serve-build": "pnpm build && pnpm http-server dist -p 8080 -P http://localhost:8080? -c-1"
     },
     "devDependencies": {
-        "@swc/core": "1.3.85",
+        "@swc/core": "1.3.86",
         "@swc/helpers": "0.5.2",
-        "@types/react": "18.2.21",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
         "@types/webpack": "5.28.2",
         "@workleap/browserslist-config": "2.0.0",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/swc-configs": "2.1.2",
         "@workleap/typescript-configs": "3.0.2",
-        "@workleap/webpack-configs": "1.0.7",
+        "@workleap/webpack-configs": "1.0.8",
         "browserslist": "4.21.10",
         "copyfiles": "2.4.1",
         "http-server": "14.1.1",
