+++ sample/host/tsconfig.json
@@ -5,7 +5,7 @@
             "@squide/core": ["../../packages/core/src/index.ts"],
             "@squide/react-router": ["../../packages/react-router/src/index.ts"],
             "@squide/webpack-module-federation": ["../../packages/webpack-module-federation/src/index.ts"],
-            "@squide/webpack-module-federation/configTransformer.js": ["../../packages/webpack-module-federation/src/configTransformer.ts"],
+            "@squide/webpack-module-federation/defineConfig.js": ["../../packages/webpack-module-federation/src/defineConfig.js"],
             "@squide/fakes": ["../../packages/fakes/src/index.ts"],
             "@sample/shell": ["../shell/src/index.ts"],
             "@sample/shared": ["../shared/src/index.ts"],
