+++ sample/local-module/package.json
@@ -25,9 +25,9 @@
         "react-router-dom": "*"
     },
     "devDependencies": {
-        "@swc/core": "1.3.85",
+        "@swc/core": "1.3.86",
         "@swc/helpers": "0.5.2",
-        "@types/react": "18.2.21",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
         "@types/webpack": "5.28.2",
         "@workleap/browserslist-config": "2.0.0",
@@ -35,7 +35,7 @@
         "@workleap/swc-configs": "2.1.2",
         "@workleap/tsup-configs": "3.0.1",
         "@workleap/typescript-configs": "3.0.2",
-        "@workleap/webpack-configs": "1.0.7",
+        "@workleap/webpack-configs": "1.0.8",
         "browserslist": "4.21.10",
         "cross-env": "7.0.3",
         "nodemon": "3.0.1",
