+++ sample/local-module/src/dev/index.tsx
@@ -1,9 +1,10 @@
 import { ConsoleLogger, Runtime, RuntimeContext, registerLocalModules } from "@squide/react-router";
-import { StrictMode, Suspense } from "react";
+import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+import { register } from "../register.tsx";
+import { sessionAccessor } from "../session.ts";
 import { App } from "./App.tsx";
-import { register } from "./register.tsx";
-import { sessionAccessor } from "./session.ts";
+import { registerTabsPage } from "./registerTabsPage.tsx";
 
 // Create the shell runtime.
 // Services, loggers and sessionAccessor could be reuse through a shared packages or faked when in isolation.
@@ -14,16 +15,14 @@ const runtime = new Runtime({
 
 // Registering the remote module as a static module because the "register" function
 // is local when developing in isolation.
-registerLocalModules([register], runtime);
+registerLocalModules([register, registerTabsPage], runtime);
 
 const root = createRoot(document.getElementById("root")!);
 
 root.render(
     <StrictMode>
         <RuntimeContext.Provider value={runtime}>
-            <Suspense fallback={<div>Loading...</div>}>
-                <App />
-            </Suspense>
+            <App />
         </RuntimeContext.Provider>
     </StrictMode>
 );
