+++ sample/remote-module/src/dev/DevHome.tsx
@@ -1,0 +1,8 @@
+export default function DevHome() {
+    return (
+        <div>
+            <h2>Remote module development home page</h2>
+            <p>Hey!</p>
+        </div>
+    );
+}
