+++ sample/remote-module/src/dev/registerTabsPage.tsx
@@ -1,10 +1,9 @@
-import type { ModuleRegisterFunction, Runtime } from "@squide/react-router";
-
+import { ModuleRegisterFunction, Runtime } from "@squide/react-router";
 import { lazy } from "react";
 
 const FederatedTabsLayout = lazy(() => import("@sample/shared/FederatedTabsLayout.tsx"));
 
-export const register: ModuleRegisterFunction<Runtime> = runtime => {
+export const registerTabsPage: ModuleRegisterFunction<Runtime> = runtime => {
     runtime.registerRoutes([
         {
             path: "/federated-tabs",
