+++ sample/remote-module/tsconfig.json
@@ -5,9 +5,10 @@
             "@squide/core": ["../../packages/core/src/index.ts"],
             "@squide/react-router": ["../../packages/react-router/src/index.ts"],
             "@squide/webpack-module-federation": ["../../packages/webpack-module-federation/src/index.ts"],
-            "@squide/webpack-module-federation/configTransformer.js": ["../../packages/webpack-module-federation/src/configTransformer.ts"],
+            "@squide/webpack-module-federation/defineConfig.js": ["../../packages/webpack-module-federation/src/defineConfig.js"],
             "@sample/shell": ["../shell/src/index.ts"],
-            "@sample/shared": ["../shared/src/index.ts"]
+            "@sample/shared": ["../shared/src/index.ts"],
+            "@sample/shared/FederatedTabsLayout.tsx": ["../shared/src/FederatedTabsLayout.tsx"]
         }
     },
     "exclude": ["dist", "node_modules"]
