+++ sample/shared/src/FederatedTabsLayout.tsx
