+++ sample/shell/package.json
@@ -29,7 +29,7 @@
         "@remix-run/router": "1.9.0",
         "@sample/shared": "workspace:*",
         "@squide/react-router": "workspace:*",
-        "@types/react": "18.2.21",
+        "@types/react": "18.2.22",
         "@types/react-dom": "18.2.7",
         "@workleap/eslint-plugin": "2.1.1",
         "@workleap/tsup-configs": "3.0.1",
